refactor: перенести все сервисы в папку services
Jira issue: VEGA-515
