+++ src/app-context.tsx
@@ -1,10 +1,10 @@
 import { createContext, useContext } from 'react';
 import { createBrowserHistory, History } from 'history';
 
-import { createGraphqlClient, GraphQLClient, ServerError } from './utils/graphql-client';
-import { Identity } from './utils/identity';
-import { Notifications } from './utils/notifications';
-import { BrowserMessageBus } from './message-bus';
+import { createGraphqlClient, GraphQLClient, ServerError } from './services/graphql-client';
+import { Identity } from './services/identity';
+import { BrowserMessageBus } from './services/message-bus';
+import { Notifications } from './services/notifications';
 
 export type ServerErrorListener = (error: ServerError) => void;
